+++ src/silhouette/__init__.py
@@ -3,6 +3,7 @@ from importlib.metadata import version
 from silhouette.two_parameter import TwoParameterRegressor
 from silhouette.three_parameter import ThreeParameterRegressor
 from silhouette.omni import OmniDurationRegressor
+from silhouette.fpca import FPCARegressor
 
 __version__ = version("silhouette")
 
@@ -10,4 +11,5 @@ __all__ = [
     "TwoParameterRegressor",
     "ThreeParameterRegressor",
     "OmniDurationRegressor",
+    "FPCARegressor",
 ]